requirements.txt
nats-py
protobuf
asyncio